migrate individual functions and test for generateFilterOptions.js to Filter.js componeent
According to colocation principles, generateFilterOption function and its test were migrated to Filter.js file as those were only used there. Individual external function and test files were removed.
